feat: adjust ViewModel usage to use consumer and select on home screen
